Move behavior to NPC package
