+++ gameserver/src/main/java/brainwine/gameserver/entity/npc/behavior/parts/IdleBehavior.java
@@ -1,4 +1,4 @@
-package brainwine.gameserver.behavior.parts;
+package brainwine.gameserver.entity.npc.behavior.parts;
 
 import java.util.Random;
 import java.util.concurrent.ThreadLocalRandom;
@@ -7,9 +7,9 @@ import com.fasterxml.jackson.annotation.JacksonInject;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonSetter;
 
-import brainwine.gameserver.behavior.Behavior;
 import brainwine.gameserver.entity.FacingDirection;
 import brainwine.gameserver.entity.npc.Npc;
+import brainwine.gameserver.entity.npc.behavior.Behavior;
 import brainwine.gameserver.util.Vector2i;
 
 public class IdleBehavior extends Behavior {
