+++ gameserver/src/main/java/brainwine/gameserver/entity/npc/behavior/parts/ReporterBehavior.java
@@ -1,10 +1,10 @@
-package brainwine.gameserver.behavior.parts;
+package brainwine.gameserver.entity.npc.behavior.parts;
 
 import com.fasterxml.jackson.annotation.JacksonInject;
 import com.fasterxml.jackson.annotation.JsonCreator;
 
-import brainwine.gameserver.behavior.Behavior;
 import brainwine.gameserver.entity.npc.Npc;
+import brainwine.gameserver.entity.npc.behavior.Behavior;
 
 public class ReporterBehavior extends Behavior {
