+++ gameserver/src/main/java/brainwine/gameserver/entity/npc/behavior/parts/UnblockBehavior.java
@@ -1,4 +1,4 @@
-package brainwine.gameserver.behavior.parts;
+package brainwine.gameserver.entity.npc.behavior.parts;
 
 import java.util.Random;
 import java.util.concurrent.ThreadLocalRandom;
@@ -6,8 +6,8 @@ import java.util.concurrent.ThreadLocalRandom;
 import com.fasterxml.jackson.annotation.JacksonInject;
 import com.fasterxml.jackson.annotation.JsonCreator;
 
-import brainwine.gameserver.behavior.Behavior;
 import brainwine.gameserver.entity.npc.Npc;
+import brainwine.gameserver.entity.npc.behavior.Behavior;
 import brainwine.gameserver.util.Vector2i;
 import brainwine.gameserver.zone.Zone;
